+++ crates/core/view.rs
@@ -4,21 +4,9 @@
 
 // module
 use crossbeam_channel::{Receiver, Sender};
-use crossterm::{
-    event::{DisableMouseCapture, EnableMouseCapture},
-    execute,
-    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
-};
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
-use std::{error::Error, io};
-use tui::{backend::CrosstermBackend, style::Color, Terminal};
-
-// non blocking io
-#[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-use nix::fcntl::{fcntl, FcntlArg::*, OFlag};
-#[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-use std::io::stdin;
+use std::error::Error;
+use tui::style::Color;
 
 // local module
 use crate::app::App;
@@ -226,30 +214,11 @@ impl View {
         rx: Receiver<AppEvent>,
         exist_results: Vec<CommandResult>,
     ) -> Result<(), Box<dyn Error>> {
-        // Setup Terminal
-        enable_raw_mode()?;
-
-        let mut stdout = io::stdout();
-        execute!(stdout, EnterAlternateScreen)?;
-
-        let backend = CrosstermBackend::new(stdout);
-        let mut terminal: Terminal<CrosstermBackend<io::Stdout>> = Terminal::new(backend)?;
-        let _ = terminal.clear();
+        let mut terminal = setup_terminal()?;
 
         {
             let input_tx = tx.clone();
-            let _ = std::thread::spawn(move || {
-                // non blocking io
-                #[cfg(any(
-                    target_os = "freebsd",
-                    target_os = "linux",
-                    target_os = "macos",
-                    target_os = "windows"
-                ))]
-                loop {
-                    let _ = send_input(input_tx.clone());
-                }
-            });
+            spawn_input_thread(input_tx);
         }
 
         // Create App
@@ -261,66 +230,7 @@ impl View {
             self.diff_mode_width,
         );
 
-        // set keymap
-        app.set_keymap(self.keymap.clone());
-
-        // set after command
-        app.set_after_command(self.after_command.clone());
-        app.set_after_command_result_write_file(self.after_command_result_write_file);
-
-        // set watch diff highlight colors
-        app.set_watch_diff_colors(self.watch_diff_fg, self.watch_diff_bg);
-
-        // set mouse events
-        // Windows mouse capture implemention requires EnableMouseCapture be invoked before DisableMouseCapture can be used
-        // https://github.com/crossterm-rs/crossterm/issues/660
-        if cfg!(target_os = "windows") {
-            execute!(terminal.backend_mut(), EnableMouseCapture)?;
-        }
-        app.set_mouse_events(self.mouse_events);
-
-        // set limit
-        app.set_limit(self.limit);
-
-        // set beep
-        app.set_beep(self.beep);
-        app.set_exit_on_change(self.exit_on_change);
-
-        // set border
-        app.set_border(self.border);
-        app.set_scroll_bar(self.scroll_bar);
-
-        // set logfile path.
-        app.set_logpath(self.log_path.clone());
-
-        // set color
-        app.set_ansi_color(self.color);
-
-        app.show_history(self.show_ui);
-        app.show_ui(self.show_ui);
-        app.show_help_banner(self.show_help_banner);
-
-        app.set_tab_size(self.tab_size);
-
-        // set line_number
-        app.set_line_number(self.line_number);
-
-        // set reverse
-        app.set_reverse(self.reverse);
-
-        // set wrap mode
-        app.set_wrap_mode(self.wrap);
-
-        // set output mode
-        app.set_output_mode(self.output_mode);
-
-        // set diff mode
-        app.set_diff_mode(self.diff_mode);
-        app.set_is_only_diffline(self.is_only_diffline);
-        app.set_ignore_spaceblock(self.ignore_spaceblock);
-
-        // set enable summary char
-        app.set_enable_summary_char(self.enable_summary_char);
+        self.apply_to_app(&mut app, &mut terminal)?;
 
         // set exist results
         app.add_results(exist_results);
@@ -337,94 +247,6 @@ impl View {
         Ok(())
     }
 }
-
-fn restore_terminal(terminal: &mut Terminal<CrosstermBackend<io::Stdout>>) {
-    let _ = disable_raw_mode();
-    let _ = execute!(
-        terminal.backend_mut(),
-        LeaveAlternateScreen,
-        DisableMouseCapture
-    );
-    let _ = terminal.show_cursor();
-}
-
-fn send_input(tx: Sender<AppEvent>) -> io::Result<()> {
-    if crossterm::event::poll(Duration::from_millis(100))? {
-        // On macOS, mouse scroll sequences can be split in a way that makes
-        // blocking stdin reads mis-handle the pending bytes. Keep stdin
-        // temporarily non-blocking only around `event::read()` so scroll input
-        // is consumed without reintroducing the old os error 35 exit path.
-        #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-        set_nonblocking(true)?;
-
-        // get event
-        let result = crossterm::event::read();
-
-        // blocking mode
-        #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-        set_nonblocking(false)?;
-
-        match result {
-            Ok(event) => {
-                let _ = tx.send(AppEvent::TerminalEvent(event));
-            }
-            Err(err)
-                if matches!(
-                    err.kind(),
-                    io::ErrorKind::WouldBlock | io::ErrorKind::Interrupted
-                ) => {}
-            Err(err) => return Err(err),
-        }
-    }
-    Ok(())
-}
-
-#[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-pub fn set_nonblocking(is_nonblocking: bool) -> nix::Result<()> {
-    let stdin = stdin();
-    let flags = fcntl(&stdin, F_GETFL)?;
-    let new_flags = next_nonblocking_flags(flags, is_nonblocking);
-    fcntl(&stdin, F_SETFL(new_flags))?;
-
-    Ok(())
-}
-
-#[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-fn next_nonblocking_flags(current_flags: i32, is_nonblocking: bool) -> OFlag {
-    let current_flags = OFlag::from_bits_truncate(current_flags);
-
-    if is_nonblocking {
-        current_flags | OFlag::O_NONBLOCK
-    } else {
-        current_flags & !OFlag::O_NONBLOCK
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-    use super::next_nonblocking_flags;
-    #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-    use nix::fcntl::OFlag;
-
-    #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-    #[test]
-    fn next_nonblocking_flags_enables_nonblocking_without_dropping_other_flags() {
-        let current_flags = (OFlag::O_APPEND | OFlag::O_CLOEXEC).bits();
-        let next_flags = next_nonblocking_flags(current_flags, true);
-
-        assert!(next_flags.contains(OFlag::O_NONBLOCK));
-        assert!(next_flags.contains(OFlag::O_APPEND));
-        assert!(next_flags.contains(OFlag::O_CLOEXEC));
-    }
-
-    #[cfg(any(target_os = "freebsd", target_os = "linux", target_os = "macos"))]
-    #[test]
-    fn next_nonblocking_flags_disables_nonblocking_without_dropping_other_flags() {
-        let current_flags = (OFlag::O_APPEND | OFlag::O_NONBLOCK).bits();
-        let next_flags = next_nonblocking_flags(current_flags, false);
-
-        assert!(!next_flags.contains(OFlag::O_NONBLOCK));
-        assert!(next_flags.contains(OFlag::O_APPEND));
-    }
-}
+#[path = "view_runtime.rs"]
+mod runtime;
+use self::runtime::{restore_terminal, setup_terminal, spawn_input_thread};
